add redis connection pool
