+++ controller/submit.go
@@ -1,20 +1,12 @@
 package controller
 
-import (
-	"log"
-	"net/http"
-
-	"github.com/garyburd/redigo/redis"
-)
+import "net/http"
 
 //HandleSubmitCode : Handle user submited code from problem page
 func HandleSubmitCode(w http.ResponseWriter, r *http.Request) {
 	if r.Method == "POST" {
-		c, err := redis.Dial("tcp", "localhost:6379")
-		if err != nil {
-			log.Fatal(err)
-		}
+		c := RedisPool.Get()
 		defer c.Close()
-
+		//	c.Do("LPUSH", "judgeQueue", "test")
 	}
 }
