+++ templates/problem.tmpl
@@ -13,7 +13,7 @@
     <hr>
     <pre>{{.SampleOutput}}</pre>
     <h6>Code</h6>
-    <form method="post" action="#">
+    <form method="post" action="/problem/submit">
         <textarea class="form-control" rows="3">Paste code here.</textarea>
         <br>
         <button tyoe="submit" class="btn btn-default">Submit</button>
